Add files via upload
This code displays the most basic deep sleep with a timer to wake it up and how to store data in RTC memory to use it over reboots

The example disables the LoBat interrupt generated by the LC709203F to prevent excessive Deep Sleep current  when the LC709203F is not completely initialized and configured for a specific cell.  It allows rapid evaluation of the Deep Sleep current of the board.